chore: move language.ts to model dir
